feat: bring back pg_dumpall support in migrate-tool
Summary: Easier to use now anyway and unified under a single interface.

+++ tools/migrate-tool.sh
@@ -38,8 +38,16 @@ else
     exit 1
 fi
 
+# thid argument is the upgrade method: either pg_dumpall or pg_ugprade
+if [ "$3" != "pg_dumpall" ] && [ "$3" != "pg_upgrade" ]; then
+    echo "Please provide a valid upgrade method (pg_dumpall or pg_upgrade)"
+    exit 1
+fi
+UPGRADE_METHOD="$3"
+
 echo "Old server build: PSQL $1"
 echo "New server build: PSQL $2"
+echo "Upgrade method: $UPGRADE_METHOD"
 
 PORTNO="${2:-5432}"
 DATDIR=$(mktemp -d)
@@ -48,7 +56,7 @@ mkdir -p "$DATDIR" "$NEWDAT"
 
 echo "NOTE: using temporary directory $DATDIR for PSQL $1 data, which will not be removed"
 echo "NOTE: you are free to re-use this data directory at will"
-echo 
+echo
 
 $OLDVER/bin/initdb -D "$DATDIR" --locale=C
 $NEWVER/bin/initdb -D "$NEWDAT" --locale=C
@@ -63,7 +71,7 @@ for x in "$DATDIR" "$NEWDAT"; do
     $PSQL_CONF_FILE > "$x/postgresql.conf"
 done
 
-echo "NOTE: Starting old server (v${1}) for temporary server to load data into the system"
+echo "NOTE: Starting first server (v${1}) to load data into the system"
 $OLDVER/bin/pg_ctl start -D "$DATDIR"
 
 PRIMING_SCRIPT=$PWD/tests/prime.sql
@@ -72,21 +80,43 @@ MIGRATION_DATA=$PWD/tests/migrations/data.sql
 $OLDVER/bin/psql -h localhost -d postgres -Xf "$PRIMING_SCRIPT"
 $OLDVER/bin/psql -h localhost -d postgres -Xf "$MIGRATION_DATA"
 
-echo "NOTE: Stopping old server (v${1}) to prepare for migration"
-$OLDVER/bin/pg_ctl stop -D "$DATDIR"
+if [ "$UPGRADE_METHOD" == "pg_upgrade" ]; then
+  echo "NOTE: Stopping old server (v${1}) to prepare for migration"
+  $OLDVER/bin/pg_ctl stop -D "$DATDIR"
 
-echo "NOTE: Migrating old data $DATDIR to $NEWDAT using pg_upgrade"
+  echo "NOTE: Migrating old data $DATDIR to $NEWDAT using pg_upgrade"
 
-export PGDATAOLD="$DATDIR"
-export PGDATANEW="$NEWDAT"
-export PGBINOLD="$OLDVER/bin"
-export PGBINNEW="$NEWVER/bin"
+  export PGDATAOLD="$DATDIR"
+  export PGDATANEW="$NEWDAT"
+  export PGBINOLD="$OLDVER/bin"
+  export PGBINNEW="$NEWVER/bin"
 
-if ! $NEWVER/bin/pg_upgrade --check; then
-    echo "ERROR: pg_upgrade check failed"
-    exit 1
+  if ! $NEWVER/bin/pg_upgrade --check; then
+      echo "ERROR: pg_upgrade check failed"
+      exit 1
+  fi
+
+  echo "NOTE: pg_upgrade check passed, proceeding with migration"
+  $NEWVER/bin/pg_upgrade
+  rm -f delete_old_cluster.sh # we don't need this
+  exit 0
 fi
 
-echo "NOTE: pg_upgrade check passed, proceeding with migration"
-$NEWVER/bin/pg_upgrade
-rm -f delete_old_cluster.sh # we don't need this
+if [ "$UPGRADE_METHOD" == "pg_dumpall" ]; then
+    SQLDAT="$DATDIR/dump.sql"
+    echo "NOTE: Exporting data via pg_dumpall ($SQLDAT)"
+    $OLDVER/bin/pg_dumpall -h localhost > "$SQLDAT"
+
+    echo "NOTE: Stopping old server (v${1}) to prepare for migration"
+    $OLDVER/bin/pg_ctl stop -D "$DATDIR"
+
+    echo "NOTE: Starting second server (v${2}) to load data into the system"
+    $NEWVER/bin/pg_ctl start -D "$NEWDAT"
+
+    echo "NOTE: Loading data into new server (v${2}) via 'cat | psql'"
+    cat "$SQLDAT" | $NEWVER/bin/psql -h localhost -d postgres
+
+    printf "\n\n\n\n"
+    echo "NOTE: Done, check logs. Stopping the server; new database is located at $NEWDAT" 
+    $NEWVER/bin/pg_ctl stop -D "$NEWDAT"
+fi
